Add blockchain and retrieval keys to settings. Separate internal and public settings representation (#14)
Internal and public settings have been separated. The public settings interface is in the fcrclient module. The implementation, which is the object passed around inside the client, is in internal/settinngs.

Blockchain private keys and algorithms plus Retrieval keys, algorithms, and key versions are now supported by the settings. The keys are not yet used.