gopkg scripts + fixes to crypto usage (#15)
This PR adds scripts to help ensure go.mod file in client, that relies on the gateway, is correctly configured.

This PR forced an upgrade to the latest version of the gateway. To ensure code compiled, the client code was changed so that it could use the latest crypto code.